==> notebooks/01_exploracao_pandas.ipynb <==
import pandas as pd

df = pd.read_csv("../data/salaries.csv")
     

df.head()

df.info()

df.describe()

df.shape

linhas, colunas = df.shape[0], df.shape[1]
print(f"Linhas: {linhas}")
print(f"Colunas: {colunas}")

df.columns

renomear_colunas = {
    'work_year': 'ano',
    'experience_level': 'senioridade',
    'employment_type': 'contrato',
    'job_title': 'cargo',
    'salary': 'salario',
    'salary_currency': 'moeda',
    'salary_in_usd': 'usd',
    'employee_residence': 'residencia',
    'remote_ratio': 'remoto',
    'company_location': 'empresa',
    'company_size': 'tamanho_empresa'
}
df.rename(columns= renomear_colunas, inplace= True)
df.columns

df["senioridade"].value_counts()

df['contrato'].value_counts()

df['remoto'].value_counts()

df['tamanho_empresa'].value_counts()

senioridade = { 
    'SE': 'Senior',
    'MI': 'Pleno',
    'EN': 'Junior',
    'EX': 'Executivo'
}

df["senioridade"] = df["senioridade"].replace(senioridade)
df["senioridade"].value_counts()

contrato = { 
    'FT': 'Tempo Integral',
    'PT': 'Tempo Parcial',
    'FL': 'Freela',
    'CT': 'Contrato'
}

df["contrato"] = df["contrato"].replace(contrato)
df["contrato"].value_counts()

tamanho_empresa = {
    'S': 'Pequena',
    'M': 'Media',
    'L': 'Grande'
}

df["tamanho_empresa"] = df["tamanho_empresa"].replace(tamanho_empresa)
df["tamanho_empresa"].value_counts()

remoto = {
    0: 'Presencial',
    50: 'Hibrido',
    100: 'Remoto'
}

df["remoto"] = df["remoto"].replace(remoto)
df["remoto"].value_counts()

df.head()

df.describe(include="object")

df.describe()

df.to_csv('../data/salaries_aula_1_processado.csv', index=False)
print("Dados exportados com sucesso para a Aula 2!")
