==> notebooks/02_limpeza_tratamento.ipynb <==
import pandas as pd

df = pd.read_csv('../data/salaries_aula_1_processado.csv')

df.head()

# Verifica se existe valores nulos (Mostra as 5 primeiras e últimas linhas do arquivo)
df.isnull()

# Soma a quantidade de linhas com valor NULO e mostra por coluna o total
df.isnull().sum()

# Mostra cada valor único de uma determinada coluna
df["ano"].unique()

# Mostra apenas as linhas que tem a informação NULL
df[df.isnull().any(axis=1)]

import numpy as np

# Criação de um DataFrame de exemplo
df_salarios = pd.DataFrame({
    "nome": ["Any", "Felipe", "Sergio", "Gabriela", "Jussara"],
    "salario": [4000, np.nan, 5000, np.nan, 100000]
})

# Substitue os NULOS pelas médias com arredondamento
df_salarios["salario_media"] = df_salarios["salario"].fillna(df_salarios["salario"].mean().round(2))

# Substitue os NULOS pela mediana dos salários
df_salarios["salario_mediana"] = df_salarios["salario"].fillna(df_salarios["salario"].median())

df_salarios

df_temperatura = pd.DataFrame({
    "dia": ["segunda", "terca", "quarta", "quinta", "sexta"],
    "temperatura": [30, np.nan, np.nan, 28, 27]
})

df_temperatura["preenchido_ffill"] = df_temperatura["temperatura"].ffill()

df_temperatura["preenchido_bfill"] = df_temperatura["temperatura"].bfill()


df_temperatura

df_cidades = pd.DataFrame ({
    "nome": ["Any", "Felipe", "Sergio", "Gabriela", "Jussara"],
    "cidade": ["São Paulo", np.nan, "Sorocaba", np.nan, "Campinas"]
})

df_cidades["cidade_preenchida"] = df_cidades["cidade"].fillna("Não informado")

df_cidades

df_limpo = df.dropna()

df_limpo.isnull().sum()

df_limpo.head()

df_limpo.info()

df_limpo = df_limpo.assign(ano= df_limpo["ano"].astype("int64"))
df_limpo

# Salvar o arquivo processado desse notebook
df_limpo.to_csv('../data/salaries_aula_2_processado.csv', index=False)
print("Checkpoint da Aula 02 salvo com sucesso!")
